separated code into modules and other improvements
